+++ algosdk/abi/contract.py
@@ -1,5 +1,5 @@
 import json
-from typing import List, Union
+from typing import Dict, List, Union
 
 from algosdk.abi.method import Method
 
@@ -10,22 +10,32 @@ class Contract:
 
     Args:
         name (string): name of the contract
-        app_id (int): application id associated with the contract
         methods (list): list of Method objects
+        desc (string, optional): description of the contract
+        networks (dict, optional): information about the contract in a
+            particular network, such as an app-id.
     """
 
-    def __init__(self, name: str, app_id: int, methods: List[Method]) -> None:
+    def __init__(
+        self,
+        name: str,
+        methods: List[Method],
+        desc: str = None,
+        networks: Dict[str, "NetworkInfo"] = None,
+    ) -> None:
         self.name = name
-        self.app_id = int(app_id)
         self.methods = methods
+        self.desc = desc
+        self.networks = networks if networks else {}
 
     def __eq__(self, o: object) -> bool:
         if not isinstance(o, Contract):
             return False
         return (
             self.name == o.name
-            and self.app_id == o.app_id
             and self.methods == o.methods
+            and self.desc == o.desc
+            and self.networks == o.networks
         )
 
     @staticmethod
@@ -36,13 +46,43 @@ class Contract:
     def dictify(self) -> dict:
         d = {}
         d["name"] = self.name
-        d["appId"] = self.app_id
         d["methods"] = [m.dictify() for m in self.methods]
+        d["desc"] = self.desc
+        d["networks"] = {k: v.dictify() for k, v in self.networks.items()}
         return d
 
     @staticmethod
     def undictify(d: dict) -> "Contract":
         name = d["name"]
-        app_id = d["appId"]
         method_list = [Method.undictify(method) for method in d["methods"]]
-        return Contract(name=name, app_id=app_id, methods=method_list)
+        desc = d["desc"] if "desc" in d else None
+        networks = d["networks"] if "networks" in d else {}
+        for k, v in networks.items():
+            networks[k] = NetworkInfo.undictify(v)
+        return Contract(
+            name=name, desc=desc, networks=networks, methods=method_list
+        )
+
+
+class NetworkInfo:
+    """
+    Represents network information.
+
+    Args:
+        app_id (int): application ID on a particular network
+    """
+
+    def __init__(self, app_id: int) -> None:
+        self.app_id = app_id
+
+    def __eq__(self, o: object) -> bool:
+        if not isinstance(o, NetworkInfo):
+            return False
+        return self.app_id == o.app_id
+
+    def dictify(self) -> dict:
+        return {"appID": self.app_id}
+
+    @staticmethod
+    def undictify(d: dict) -> "NetworkInfo":
+        return NetworkInfo(app_id=d["appID"])