+++ algosdk/atomic_transaction_composer.py
@@ -2,15 +2,17 @@ from abc import ABC, abstractmethod
 import base64
 import copy
 from enum import IntEnum
-from typing import Any, List, Union
+from typing import Any, List, TypeVar, Union
 
 from algosdk import abi, error
-from algosdk.abi.base_type import ABIType
+from algosdk.abi.address_type import AddressType
 from algosdk.future import transaction
 from algosdk.v2client import algod
 
 # The first four bytes of an ABI method call return must have this hash
 ABI_RETURN_HASH = b"\x15\x1f\x7c\x75"
+# Support for generic typing
+T = TypeVar("T")
 
 
 class AtomicTransactionComposerStatus(IntEnum):
@@ -34,6 +36,38 @@ class AtomicTransactionComposerStatus(IntEnum):
     COMMITTED = 4
 
 
+def populate_foreign_array(
+    value_to_add: T, foreign_array: List[T], zero_value: T = None
+) -> int:
+    """
+    Add a value to an application call's foreign array. The addition will be as
+    compact as possible, and this function will return an index used to
+    reference `value_to_add` in the `foreign_array`.
+
+    Args:
+        value_to_add: value to add to the array. If the value is already
+            present, it will not be added again. Instead, the existing index
+            will be returned.
+        foreign_array: the existing foreign array. This input may be modified
+            to append `value_to_add`.
+        zero_value: If provided, this value indicates two things: the 0 value is
+            reserved for this array so `foreign_array` must start at index 1;
+            additionally, if `value_to_add` equals `zero_value`, then
+            `value_to_add` will not be added to the array and the 0 index will
+            be returned.
+    """
+    if zero_value and value_to_add == zero_value:
+        return 0
+
+    offset = 0 if not zero_value else 1
+
+    if value_to_add in foreign_array:
+        return foreign_array.index(value_to_add) + offset
+
+    foreign_array.append(value_to_add)
+    return offset + len(foreign_array) - 1
+
+
 class AtomicTransactionComposer:
     """
     Constructs an atomic transaction group which may contain a combination of
@@ -50,7 +84,7 @@ class AtomicTransactionComposer:
     # The maximum size of an atomic transaction group.
     MAX_GROUP_SIZE = 16
     # The maximum number of app-args that can be individually packed for ABIs
-    MAX_ABI_APP_ARG_LIMIT = 14
+    MAX_APP_ARG_LIMIT = 16
 
     def __init__(self) -> None:
         self.status = AtomicTransactionComposerStatus.BUILDING
@@ -122,12 +156,20 @@ class AtomicTransactionComposer:
     def add_method_call(
         self,
         app_id: int,
-        method: abi.method.Method,
+        method: abi.Method,
         sender: str,
         sp: transaction.SuggestedParams,
         signer: "TransactionSigner",
-        method_args: List[Union[Any]] = None,
+        method_args: List[Union[Any, "TransactionWithSigner"]] = None,
         on_complete: transaction.OnComplete = transaction.OnComplete.NoOpOC,
+        local_schema: transaction.StateSchema = None,
+        global_schema: transaction.StateSchema = None,
+        approval_program: bytes = None,
+        clear_program: bytes = None,
+        extra_pages: int = None,
+        accounts: List[str] = None,
+        foreign_apps: List[int] = None,
+        foreign_assets: List[int] = None,
         note: bytes = None,
         lease: bytes = None,
         rekey_to: str = None,
@@ -150,6 +192,19 @@ class AtomicTransactionComposer:
                 or transactions that immediate precede this method call
             on_complete (OnComplete, optional): intEnum representing what app should do on completion
                 and if blank, it will default to a NoOp call
+            local_schema (StateSchema, optional): restricts what can be stored by created application;
+                must be omitted if not creating an application
+            global_schema (StateSchema, optional): restricts what can be stored by created application;
+                must be omitted if not creating an application
+            approval_program (bytes, optional): the program to run on transaction approval;
+                must be omitted if not creating or updating an application
+            clear_program (bytes, optional): the program to run when state is being cleared;
+                must be omitted if not creating or updating an application
+            extra_pages (int, optional): additional program space for supporting larger programs.
+                A page is 1024 bytes.
+            accounts (list[string], optional): list of additional accounts involved in call
+            foreign_apps (list[int], optional): list of other applications (identified by index) involved in call
+            foreign_assets (list[int], optional): list of assets involved in call
             note (bytes, optional): arbitrary optional bytes
             lease (byte[32], optional): specifies a lease, and no other transaction
                 with the same sender and lease can be confirmed in this
@@ -165,56 +220,117 @@ class AtomicTransactionComposer:
             raise error.AtomicTransactionComposerError(
                 "AtomicTransactionComposer cannot exceed MAX_GROUP_SIZE transactions"
             )
+        if app_id == 0:
+            if not approval_program or not clear_program:
+                raise error.AtomicTransactionComposerError(
+                    "One of the following required parameters for application creation is missing: approvalProgram, clearProgram"
+                )
+        elif on_complete == transaction.OnComplete.UpdateApplicationOC:
+            if not approval_program or not clear_program:
+                raise error.AtomicTransactionComposerError(
+                    "One of the following required parameters for OnApplicationComplete.UpdateApplicationOC is missing: approvalProgram, clearProgram"
+                )
+            if local_schema or global_schema or extra_pages:
+                raise error.AtomicTransactionComposerError(
+                    "One of the following application creation parameters were set on a non-creation call: numGlobalInts, numGlobalByteSlices, numLocalInts, numLocalByteSlices, extraPages"
+                )
+        elif (
+            approval_program
+            or clear_program
+            or local_schema
+            or global_schema
+            or extra_pages
+        ):
+            raise error.AtomicTransactionComposerError(
+                "One of the following application creation parameters were set on a non-creation call: approvalProgram, clearProgram, numGlobalInts, numGlobalByteSlices, numLocalInts, numLocalByteSlices, extraPages"
+            )
         if not method_args:
             method_args = []
         if len(method.args) != len(method_args):
             raise error.AtomicTransactionComposerError(
                 "number of method arguments do not match the method signature"
             )
-        if not isinstance(method, abi.method.Method):
+        if not isinstance(method, abi.Method):
             raise error.AtomicTransactionComposerError(
                 "invalid Method object was passed into AtomicTransactionComposer"
             )
-        if app_id == 0:
-            raise error.AtomicTransactionComposerError(
-                "application create call not supported"
-            )
+
+        # Initialize foreign object maps
+        accounts = accounts[:] if accounts else []
+        foreign_apps = foreign_apps[:] if foreign_apps else []
+        foreign_assets = foreign_assets[:] if foreign_assets else []
 
         app_args = []
-        # For more than 14 args, including the selector, compact them into a tuple
-        additional_args = []
-        additional_types = []
+        raw_values = []
+        raw_types = []
         txn_list = []
+
         # First app arg must be the selector of the method
         app_args.append(method.get_selector())
+
         # Iterate through the method arguments and either pack a transaction
         # or encode a ABI value.
         for i, arg in enumerate(method.args):
-            if arg.type in abi.method.TRANSACTION_ARGS:
-                if not isinstance(method_args[i], TransactionWithSigner):
+            if abi.is_abi_transaction_type(arg.type):
+                if not isinstance(
+                    method_args[i], TransactionWithSigner
+                ) or not abi.check_abi_transaction_type(
+                    arg.type, method_args[i].txn
+                ):
                     raise error.AtomicTransactionComposerError(
                         "expected TransactionWithSigner as method argument, but received: {}".format(
                             method_args[i]
                         )
                     )
-                if method_args[i].txn.group and method_args[i].txn.group != 0:
-                    raise error.AtomicTransactionComposerError(
-                        "cannot add a transaction with nonzero group ID"
-                    )
                 txn_list.append(method_args[i])
-            elif len(app_args) > self.MAX_ABI_APP_ARG_LIMIT:
-                # Pack the remaining values as a tuple
-                additional_types.append(arg.type)
-                additional_args.append(method_args[i])
             else:
-                encoded_arg = arg.type.encode(method_args[i])
-                app_args.append(encoded_arg)
+                if abi.is_abi_reference_type(arg.type):
+                    current_type = abi.UintType(8)
+                    if arg.type == abi.ABIReferenceType.ACCOUNT:
+                        address_type = AddressType()
+                        account_arg = address_type.decode(
+                            address_type.encode(method_args[i])
+                        )
+                        current_arg = populate_foreign_array(
+                            account_arg, accounts, sender
+                        )
+                    elif arg.type == abi.ABIReferenceType.ASSET:
+                        asset_arg = int(method_args[i])
+                        current_arg = populate_foreign_array(
+                            asset_arg, foreign_assets
+                        )
+                    elif arg.type == abi.ABIReferenceType.APPLICATION:
+                        app_arg = int(method_args[i])
+                        current_arg = populate_foreign_array(
+                            app_arg, foreign_apps, app_id
+                        )
+                    else:
+                        # Shouldn't reach this line unless someone accidentally
+                        # adds another foreign array arg
+                        raise error.AtomicTransactionComposerError(
+                            "cannot recognize {} as a foreign array arg".format(
+                                arg.type
+                            )
+                        )
+                else:
+                    current_type = arg.type
+                    current_arg = method_args[i]
 
-        if additional_args:
-            remainder_args = abi.TupleType(additional_types).encode(
-                additional_args
-            )
-            app_args.append(remainder_args)
+                raw_types.append(current_type)
+                raw_values.append(current_arg)
+
+        # Compact the arguments into a single tuple, if there are more than
+        # 15 arguments excluding the selector, into the last app arg slot.
+        if len(raw_types) > self.MAX_APP_ARG_LIMIT - 1:
+            additional_types = raw_types[self.MAX_APP_ARG_LIMIT - 2 :]
+            additional_values = raw_values[self.MAX_APP_ARG_LIMIT - 2 :]
+            raw_types = raw_types[: self.MAX_APP_ARG_LIMIT - 2]
+            raw_values = raw_values[: self.MAX_APP_ARG_LIMIT - 2]
+            raw_types.append(abi.TupleType(additional_types))
+            raw_values.append(additional_values)
+
+        for i, arg_type in enumerate(raw_types):
+            app_args.append(arg_type.encode(raw_values[i]))
 
         # Create a method call transaction
         method_txn = transaction.ApplicationCallTxn(
@@ -222,10 +338,18 @@ class AtomicTransactionComposer:
             sp=sp,
             index=app_id,
             on_complete=on_complete,
+            local_schema=local_schema,
+            global_schema=global_schema,
+            approval_program=approval_program,
+            clear_program=clear_program,
             app_args=app_args,
+            accounts=accounts,
+            foreign_apps=foreign_apps,
+            foreign_assets=foreign_assets,
             note=note,
             lease=lease,
             rekey_to=rekey_to,
+            extra_pages=extra_pages,
         )
         txn_with_signer = TransactionWithSigner(method_txn, signer)
         txn_list.append(txn_with_signer)
@@ -250,11 +374,14 @@ class AtomicTransactionComposer:
             )
 
         # Get group transaction id
-        group_txns = [t.txn for t in self.txn_list]
-        group_id = transaction.calculate_group_id(group_txns)
-        for t in self.txn_list:
-            t.txn.group = group_id
-            self.tx_ids.append(t.txn.get_txid())
+        if len(self.txn_list) > 1:
+            group_txns = [t.txn for t in self.txn_list]
+            group_id = transaction.calculate_group_id(group_txns)
+            for t in self.txn_list:
+                t.txn.group = group_id
+                self.tx_ids.append(t.txn.get_txid())
+        else:
+            self.tx_ids.append(self.txn_list[0].txn.get_txid())
 
         self.status = AtomicTransactionComposerStatus.BUILT
         return self.txn_list
@@ -354,6 +481,8 @@ class AtomicTransactionComposer:
             )
 
         self.submit(client)
+        self.status = AtomicTransactionComposerStatus.SUBMITTED
+
         resp = transaction.wait_for_confirmation(
             client, self.tx_ids[0], wait_rounds
         )
@@ -388,17 +517,24 @@ class AtomicTransactionComposer:
                 logs = resp["logs"] if "logs" in resp else []
 
                 # Look for the last returned value in the log
-                for result in reversed(logs):
-                    # Check that the first four bytes is the hash of "return"
-                    result_bytes = base64.b64decode(result)
-                    if result_bytes[:4] != ABI_RETURN_HASH:
-                        continue
-                    raw_value = result_bytes[4:]
-                    return_value = self.method_dict[i].returns.type.decode(
-                        raw_value
+                if not logs:
+                    raise error.AtomicTransactionComposerError(
+                        "app call transaction did not log a return value"
                     )
-                    decode_error = None
-                    break
+                result = logs[-1]
+                # Check that the first four bytes is the hash of "return"
+                result_bytes = base64.b64decode(result)
+                if (
+                    len(result_bytes) < 4
+                    or result_bytes[:4] != ABI_RETURN_HASH
+                ):
+                    raise error.AtomicTransactionComposerError(
+                        "app call transaction did not log a return value"
+                    )
+                raw_value = result_bytes[4:]
+                return_value = self.method_dict[i].returns.type.decode(
+                    raw_value
+                )
             except Exception as e:
                 decode_error = e